+++ tests/unit/test_clean_utf_8.c
@@ -22,6 +22,7 @@
 #include "builtin_table.h"
 #include "clean_string.h"
 #include "detox_struct.h"
+#include "table.h"
 
 #include "unit_struct.h"
 
@@ -29,49 +30,139 @@
 static struct test_filename data[DATA_COUNT] = {
     // legacy
 
-    { .filename = "safe string",                    .expected = "safe string" },
-    { .filename = "\u00AE reg",                     .expected = "_reg_ reg" },
-    { .filename = "\u00A9 copy",                    .expected = "_copy_ copy" },
-    { .filename = "\u00C6 capital AE",              .expected = "AE capital AE" },
-    { .filename = "\u00DE capital thorn",           .expected = "TH capital thorn" },
-    { .filename = "\u014A capital ENG",             .expected = "NG capital ENG" },
-    { .filename = "\u0172 capital U with Ogonek",   .expected = "U capital U with Ogonek" },
+    {
+        .filename   = "safe string",
+        .expected_a = "safe string",
+        .expected_b = "safe string",
+    },
+    {
+        .filename   = "\u00AE reg",
+        .expected_a = "_reg_ reg",
+        .expected_b = "_reg_ reg",
+    },
+    {
+        .filename   = "\u00A9 copy",
+        .expected_a = "_copy_ copy",
+        .expected_b = "_copy_ copy",
+    },
+    {
+        .filename   = "\u00C6 capital AE",
+        .expected_a = "AE capital AE",
+        .expected_b = "AE capital AE",
+    },
+    {
+        .filename   = "\u00DE capital thorn",
+        .expected_a = "TH capital thorn",
+        .expected_b = "TH capital thorn",
+    },
+    {
+        .filename   = "\u014A capital ENG",
+        .expected_a = "NG capital ENG",
+        .expected_b = "NG capital ENG",
+    },
+    {
+        .filename   = "\u0172 capital U with Ogonek",
+        .expected_a = "U capital U with Ogonek",
+        .expected_b = "U capital U with Ogonek",
+    },
 
     // github issue 40
+    // + full coverage - test default translation rules
 
-    { .filename = "_ () whatever",                  .expected = "_ () whatever" },
-    { .filename = "\u4000 ignore",                  .expected = "_ ignore" },
-    { .filename = "\x7F delete",                    .expected = "_ delete" },
+    {
+        .filename   = "_ () whatever",
+        .expected_a = "_ () whatever",
+        .expected_b = "_ () whatever",
+    },
+    {
+        .filename   = "\u4000 ignore",
+        .expected_a = "_ ignore",
+        .expected_b = "\u4000 ignore",
+    },
+    {
+        .filename   = "\x7F delete",
+        .expected_a = "_ delete",
+        .expected_b = "\x7F delete",
+    },
+
     // 2 byte UTF-8 encoded 0x00 (null)
-    { .filename = "\xC0\x80 null",                  .expected = "_hidden_null_ null" },
+    {
+        .filename   = "\xC0\x80 null",
+        .expected_a = "_hidden_null_ null",
+        .expected_b = "_hidden_null_ null",
+    },
+
     // 2 byte UTF-8 encoded 0x20 (space)
-    { .filename = "\xC0\xA0 space",                 .expected = "  space" },
+    {
+        .filename   = "\xC0\xA0 space",
+        .expected_a = "  space",
+        .expected_b = "  space",
+    },
+
     // 2 byte UTF-8 encoded 0x30 (zero)
-    { .filename = "\xC0\xB0 zero",                  .expected = "0 zero" },
+    {
+        .filename   = "\xC0\xB0 zero",
+        .expected_a = "0 zero",
+        .expected_b = "0 zero",
+    },
+
     // 2 byte UTF-8 encoded 0x7E (tilde)
-    { .filename = "\xC1\xBE tilde",                 .expected = "~ tilde" },
+    {
+        .filename   = "\xC1\xBE tilde",
+        .expected_a = "~ tilde",
+        .expected_b = "~ tilde",
+    },
+
     // 2 byte UTF-8 encoded 0x7F (tilde)
-    { .filename = "\xC1\xBF delete",                .expected = "_ delete" },
+    {
+        .filename   = "\xC1\xBF delete",
+        .expected_a = "_ delete",
+        .expected_b = "\xC1\xBF delete",
+    },
 
     // github issue 33
 
-    { .filename = "\U00010348 hwair",               .expected = "hu hwair" },
+    {
+        .filename   = "\U00010348 hwair",
+        .expected_a = "hu hwair",
+        .expected_b = "hu hwair",
+    },
 };
 
 START_TEST(test_clean_utf_8)
 {
-#line 54
-    table_t *table;
+#line 125
+    table_t *table_a;
+    table_t *table_b;
     char *output;
     int i;
 
-    table = load_builtin_unicode_table();
+    table_a = load_builtin_unicode_table();
+    table_b = load_builtin_unicode_table();
+
+    table_a->default_translation = strdup("_");
+    table_b->default_translation = NULL;
+
+    // confirm that our a/b test is still valid
+    ck_assert(table_get(table_b, 0x4000) == NULL);
+    ck_assert(table_get(table_b, 0x007F) == NULL);
 
     // legacy tests
     for (i = 0; i < DATA_COUNT; i++) {
-        output = clean_utf_8(data[i].filename, table);
-        ck_assert_str_eq(output, data[i].expected);
+        output = clean_utf_8(data[i].filename, table_a);
+        ck_assert_str_eq(output, data[i].expected_a);
+
+        output = clean_utf_8(data[i].filename, table_b);
+        ck_assert_str_eq(output, data[i].expected_b);
     }
+
+}
+END_TEST
+
+START_TEST(test_clean_utf_8_null)
+{
+#line 150
+    char *output;
 
     // confirm NULL works
     output = clean_utf_8(NULL, NULL);
@@ -82,8 +173,57 @@ END_TEST
 
 START_TEST(test_clean_utf_8_missing_table)
 {
-#line 71
+#line 157
     clean_utf_8("what", NULL);
+
+}
+END_TEST
+
+START_TEST(test_clean_utf_8_invalid)
+{
+#line 160
+    table_t *table;
+    char *output;
+
+    table = load_builtin_unicode_table();
+    table->default_translation = strdup("_");
+
+    // test an invalid UTF-8 sequence
+    // the cleaner should discard the invalid sequence
+
+    output = clean_utf_8("\xC0" "blah", table);
+    ck_assert_str_eq(output, "blah");
+
+    output = clean_utf_8("blah" "\xC0", table);
+    ck_assert_str_eq(output, "blah");
+
+}
+END_TEST
+
+START_TEST(test_clean_utf_8_invalid_5_byte)
+{
+#line 176
+    table_t *table;
+    char *output;
+
+    table = load_builtin_unicode_table();
+    table->default_translation = strdup("_");
+
+    // 3-byte encoding a lower ASCII "6"
+
+    output = clean_utf_8("\xE0\x80\xB6" "blah", table);
+    ck_assert_str_eq(output, "6blah");
+
+    // 4-byte encoding a lower ASCII "6"
+
+    output = clean_utf_8("\xF0\x80\x80\xB6" "blah", table);
+    ck_assert_str_eq(output, "6blah");
+
+    // 5-byte encoding a lower ASCII "6"
+    // this should fail
+
+    output = clean_utf_8("\xF8\x80\x80\x80\xB6" "blah", table);
+    ck_assert_str_eq(output, "6blah");
 }
 END_TEST
 
@@ -96,7 +236,10 @@ int main(void)
 
     suite_add_tcase(s1, tc1_1);
     tcase_add_test(tc1_1, test_clean_utf_8);
+    tcase_add_test(tc1_1, test_clean_utf_8_null);
     tcase_add_exit_test(tc1_1, test_clean_utf_8_missing_table, 1);
+    tcase_add_test(tc1_1, test_clean_utf_8_invalid);
+    tcase_add_exit_test(tc1_1, test_clean_utf_8_invalid_5_byte, 1);
 
     srunner_run_all(sr, CK_ENV);
     nf = srunner_ntests_failed(sr);